+++ yandex/resource_yandex_compute_instance_group.go
@@ -298,6 +298,13 @@ func resourceYandexComputeInstanceGroup() *schema.Resource {
 				},
 			},
 
+			"variables": {
+				Type:     schema.TypeMap,
+				Optional: true,
+				Elem:     &schema.Schema{Type: schema.TypeString},
+				Set:      schema.HashString,
+			},
+
 			"scale_policy": {
 				Type:     schema.TypeList,
 				MaxItems: 1,
@@ -741,6 +748,11 @@ func flattenInstanceGroup(d *schema.ResourceData, instanceGroup *instancegroup.I
 		return err
 	}
 
+	variables := flattenInstanceGroupVariable(instanceGroup.GetVariables())
+	if err := d.Set("variables", variables); err != nil {
+		return err
+	}
+
 	scalePolicy, err := flattenInstanceGroupScalePolicy(instanceGroup)
 	if err != nil {
 		return err
@@ -881,6 +893,11 @@ func prepareCreateInstanceGroupRequest(d *schema.ResourceData, meta *Config) (*i
 		return nil, fmt.Errorf("Error creating 'load_balancer_spec' object of api request: %s", err)
 	}
 
+	variables, err := expandInstanceGroupVariables(d.Get("variables"))
+	if err != nil {
+		return nil, fmt.Errorf("Error creating 'variables' object of api request: %s", err)
+	}
+
 	req := &instancegroup.CreateInstanceGroupRequest{
 		FolderId:         folderID,
 		Name:             d.Get("name").(string),
@@ -893,6 +910,7 @@ func prepareCreateInstanceGroupRequest(d *schema.ResourceData, meta *Config) (*i
 		LoadBalancerSpec: loadBalancerSpec,
 		HealthChecksSpec: healthChecksSpec,
 		ServiceAccountId: d.Get("service_account_id").(string),
+		Variables:        variables,
 	}
 
 	return req, nil
@@ -934,6 +952,11 @@ func prepareUpdateInstanceGroupRequest(d *schema.ResourceData, meta *Config) (*i
 		return nil, fmt.Errorf("Error creating 'load_balancer_spec' object of api request: %s", err)
 	}
 
+	variables, err := expandInstanceGroupVariables(d.Get("variables"))
+	if err != nil {
+		return nil, fmt.Errorf("Error creating 'variables' object of api request: %s", err)
+	}
+
 	var updatePath = getStaticUpdatePath()
 
 	var instanceGroupTemplateFieldsMap = map[string]string{
@@ -960,6 +983,7 @@ func prepareUpdateInstanceGroupRequest(d *schema.ResourceData, meta *Config) (*i
 		HealthChecksSpec: healthChecksSpec,
 		ServiceAccountId: d.Get("service_account_id").(string),
 		UpdateMask:       &field_mask.FieldMask{Paths: updatePath},
+		Variables:        variables,
 	}
 
 	return req, nil
@@ -979,6 +1003,7 @@ func getStaticUpdatePath() []string {
 		"instance_template.scheduling_policy",
 		"instance_template.service_account_id",
 		"instance_template.network_settings",
+		"variables",
 		"scale_policy",
 		"deploy_policy",
 		"allocation_policy",